+++ PaymentGateway.Application/Queries/ListOfAccounts.cs
@@ -127,11 +127,11 @@ namespace PaymentGateway.Application.Queries
         public class Model
         {
             public int Id { get; set; }
-            public double Balance { get; set; }
+            public decimal Balance { get; set; }
             public string Currency { get; set; }
             public string Iban { get; set; }
             public string Status { get; set; }
-            public double Limit { get; set; }
+            public decimal Limit { get; set; }
             public string Type { get; set; }
         }
     }
