+++ PaymentGateway.Models/Product.cs
@@ -6,9 +6,9 @@ namespace PaymentGateway.Models
     {
         public int Id  { get; set; }
         public string Name { get; set; }
-        public double Value { get; set; }
+        public decimal Value { get; set; }
 
         public string Currency { get; set; }
-        public int Limit { get; set; }
+        public decimal Limit { get; set; }
     }
 }
